services/interaction_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
